+++ mt7915/mcu.c
@@ -522,7 +522,7 @@ mt7915_mcu_rx_log_message(struct mt7915_dev *dev, struct sk_buff *skb)
 	}
 
 	wiphy_info(mt76_hw(dev)->wiphy, "%s: %*s", type,
-		   skb->len - sizeof(*rxd), data);
+		   (int)(skb->len - sizeof(*rxd)), data);
 }
 
 static void
